requirements.txt
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
numpy==1.26.0
openai==1.3.0
matplotlib==3.8.0
wordcloud==1.9.2
seaborn==0.13.0
pyyaml==6.0.1
pytest==7.4.2
apache-airflow==2.7.1
jupyter==1.0.0